backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pytest
pytest-asyncio
APScheduler
pydantic-settings
respx
asgi-lifespan
python-dateutil